Error and Loading to movieList moved; Loading from Navigation deleted and enabled at the start
