+++ src/features/NoResult/index.js
@@ -1,18 +1,21 @@
+import { GlobalWrapper } from "../../common/GlobalWrapper/styled";
 import Header from "../../common/Header";
 import { StyledHeader, StyledBigCloud, StyledSmallCloud, NoResultImage, StyledPaperPlane, StyledPlaneTrack, Image } from "./styled";
 
 const NoResult = () => (
-    <StyledHeader>
-        <Header text="Sorry, there are no results for “Muan”" />
-        <NoResultImage>
-            <Image>
-                <StyledBigCloud />
-                <StyledSmallCloud />
-                <StyledPaperPlane />
-                <StyledPlaneTrack />
-            </Image>
-        </NoResultImage>
-    </StyledHeader>
+    <GlobalWrapper>
+        <StyledHeader>
+            <Header text="Sorry, there are no results for “Muan”" />
+            <NoResultImage>
+                <Image>
+                    <StyledBigCloud />
+                    <StyledSmallCloud />
+                    <StyledPaperPlane />
+                    <StyledPlaneTrack />
+                </Image>
+            </NoResultImage>
+        </StyledHeader>
+    </GlobalWrapper>
 );
 
 export default NoResult;
