+++ src/features/movieList/Error/styled.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { ReactComponent as DangerIcon } from "../../images/danger.svg";
+import { ReactComponent as DangerIcon } from "../../../images/danger.svg";
 
 export const Wrapper = styled.div`
     margin-top: 180px;
