+++ src/features/movieList/Loading/index.js
@@ -1,0 +1,16 @@
+import { StyledHeader, SpinnerSpace, StyledSpinner } from "./styled";
+import Header from "../../../common/Header";
+import { GlobalWrapper } from "../../../common/GlobalWrapper/styled";
+
+const Loading = () => (
+    <GlobalWrapper>
+        <StyledHeader>
+            <Header text="" />
+            <SpinnerSpace>
+                <StyledSpinner />
+            </SpinnerSpace>
+        </StyledHeader>
+    </GlobalWrapper>
+);
+
+export default Loading;
